moved types and components folder
